pyproject.toml
[tool.poetry]
name = "auo_project"
version = "0.1.0"
description = "AUO Health Backend Project"
authors = [
    "Luffy Tsai <luffy.tsai@val.tw>"
]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
fastapi = "^0.85.0"
fastapi-pagination = "^0.10.0"
fastapi_async_sqlalchemy = "^0.3.12"
uvicorn = { version = "^0.17.0", extras = ["standard"] }
pydantic = {extras = ["email"], version = "^1.10.2"}
yarl = "^1.7.2"
ujson = "^5.1.0"
SQLAlchemy = {version = "^1.4", extras = ["mypy", "asyncio"]}
asyncpg = {version = "^0.25.0", extras = ["sa"]}
redis = {version = "^4.3.3", extras = ["hiredis"]}
httptools = "^0.3.0"
prometheus-client = "^0.14.1"
prometheus-fastapi-instrumentator = "5.8.1"
sentry-sdk = "^1.5.12"
opentelemetry-api = {version = "^1.12.0rc2", allow-prereleases = true}
opentelemetry-sdk = {version = "^1.12.0rc2", allow-prereleases = true}
opentelemetry-exporter-otlp = {version = "^1.12.0rc2", allow-prereleases = true}
opentelemetry-instrumentation = "^0.32b0"
opentelemetry-instrumentation-fastapi = "^0.32b0"
opentelemetry-instrumentation-redis = "^0.32b0"
opentelemetry-instrumentation-sqlalchemy = "^0.32b0"
loguru = "^0.6.0"
sqlmodel = "^0.0.8"
orjson = "^3.8.0"
cryptography = "^38.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
alembic = "^1.8.1"
azure-storage-blob = "^12.13.1"
pydash = "^5.1.1"
typer = {extras = ["all"], version = "^0.6.1"}
ipython = "^8.5.0"
python-dateutil = "^2.8.2"
types-python-dateutil = "^2.8.19"
celery = "^5.2.7"
asgiref = "^3.5.2"
pytest = "^7.2.0"
httpx = "^0.22.0"
pytest-env = "^0.6.2"
fakeredis = "^1.8.1"
pytest-asyncio = "^0.19.0"
pycryptodome = "^3.16.0"
pandas = "^1.5.2"
sqlalchemy-mixins = "^1.5.3"
streamlit = "1.28"
openpyxl = "^3.1.2"
psycopg2-binary = "^2.9.9"
streamlit-code-editor = "^0.1.10"
better-exceptions = "^0.3.3"
imutils = "^0.5.4"
opencv-python-headless = "^4.9.0.80"
azure-communication-email = "^1.0.0"
pdfkit = "1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
flake8 = "^4.0.1"
mypy = "^0.961"
isort = "^5.10.1"
yesqa = "^1.3.0"
pre-commit = "^2.19.0"
wemake-python-styleguide = "^0.16.1"
black = "^22.3.0"
autoflake = "^1.4"
pytest-cov = "^3.0.0"
anyio = "^3.6.1"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["auo_project",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
